fix: New logic shared between selection and util
